Rewrote UCLLineChartView to use UIScrollView.
It now embeds a UIScrollView which has another UIView subclass as it's
scrollable view. This ChartView has separate CALayers for each line of
values added to the chart. The main UCLLineChartView is responsible for
drawing the axes and markers.

Zooming "works" but looks like crap. Need to figure out how to
counteract the image scaling but keep the scrollview working.